backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pymongo==4.7.3
python-dotenv==1.0.1
supabase==2.5.0
google-generativeai==0.7.2
pydantic==2.7.4
